+++ user_docs/en/changes.md
@@ -56,6 +56,7 @@ The several built-in table definitions are moved to the `__tables` module in tha
 
 * `NVDAHelper.versionedLibPath` is deprecated. Use `NVDAHelper.versionedLibX86Path` instead. (#18207)
 * `NVDAHelper.generateBeep` is deprecated. Use `NVDAHelper.localLib.generateBeep` instead. (#18207)
+* `NVDAHelper.VBuf_getTextInRange` is deprecated. Use `NVDAHelper.localLib.VBuf_getTextInRange` instead. (#18207)
 
 ## 2025.1.1